Fix security issue.
The docs state that you need to make the config file _readable_
and then proceeds to make the file writable.

This changes it to make the file to be owned by root and readable by
anyone. This is the default for unix / linux and suggested practice
for files in /etc.